fix: compile contracts
